backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
langchain>=0.3.0
langchain-openai>=0.3.0
openai>=1.86.0
python-dotenv>=1.0.0
pydantic>=2.7.4,<3.0.0
langfuse>=3.0.0
python-multipart>=0.0.6
httpx>=0.25.0